feat: get paginated ratings added

+++ src/GoodReads.Application/Features/Ratings/RatingResponse.cs
@@ -1,7 +1,7 @@
 namespace GoodReads.Application.Features.Ratings
 {
     public sealed record RatingResponse(
-        decimal Score,
+        int Score,
         string Description,
         ReadingResponse Reading,
         DateTime CreatedAt
